Move Pattpass test code to ConsoleExample.java
