+++ src/main/java/cl/transbank/patpass/PatpassByWebpay.java
@@ -142,67 +142,4 @@ public class PatpassByWebpay {
             }
         }
     }
-
-    public static void main(String[] args) {
-        System.setProperty("java.util.logging.SimpleFormatter.format", "[%1$tF %1$tT] [%4$-7s] %5$s %n");
-        Logger globalLog = Logger.getLogger("cl.transbank");
-        globalLog.setUseParentHandlers(false);
-        globalLog.addHandler(new ConsoleHandler() {
-            {/*setOutputStream(System.out);*/setLevel(Level.ALL);}
-        });
-        globalLog.setLevel(Level.ALL);
-
-        try {
-            String buyOrder = String.valueOf(new Random().nextInt(Integer.MAX_VALUE));
-            String sessionId = String.valueOf(new Random().nextInt(Integer.MAX_VALUE));
-            String returnUrl = "https://vuelta.com";
-            String serviceId = nextString(20);
-            String cardHolderId = nextString(20);
-            String cardHolderName = nextString(20);
-            String cardHolderLastName1 = nextString(20);
-            String cardHolderLastName2 = nextString(20);
-            String cardHolderMail = String.format("%s@%s.COM", nextString(10), nextString(7));
-            String cellphoneNumber = String.valueOf(new Random().nextInt(Integer.MAX_VALUE));
-            String expirationDate = "2222-11-11";
-            String commerceMail = String.format("%s@%s.COM", nextString(10), nextString(7));
-            final CreatePatpassByWebpayTransactionResponse response = Transaction.create(buyOrder, sessionId, 1000, returnUrl, serviceId, cardHolderId, cardHolderName,
-                    cardHolderLastName1, cardHolderLastName2, cardHolderMail, cellphoneNumber, expirationDate, commerceMail, false, null);
-            logger.info(response.toString());
-        } catch (CreateTransactionException e) {
-            e.printStackTrace();
-        }
-
-        try {
-            String token = "e82f8d3efef39449c10c66bc2f121f5af0041e57dff3cec5660e0b2be4251740";
-            final CommitPatpassByWebpayTransactionResponse response = Transaction.commit(token);
-            logger.info(response.toString());
-        } catch (Exception e) {
-            e.printStackTrace();
-        }
-
-        try {
-            String token = "e82f8d3efef39449c10c66bc2f121f5af0041e57dff3cec5660e0b2be4251740";
-            final RefundPatpassByWebpayTransactionResponse response = Transaction.refund(token, 10);
-            logger.info(response.toString());
-        } catch (RefundTransactionException e) {
-            e.printStackTrace();
-        }
-
-        try {
-            String token = "e82f8d3efef39449c10c66bc2f121f5af0041e57dff3cec5660e0b2be4251740";
-            final StatusPatpassByWebpayTransactionResponse response = Transaction.status(token);
-            logger.info(response.toString());
-        } catch (StatusTransactionException e) {
-            e.printStackTrace();
-        }
-    }
-
-    static String nextString(int length) {
-        char[] buf = new char[length];
-        Random random = new Random();
-        final char[] symbols = "ABCDEFGHIJKLMNOPQRSTUVWXYZ".toCharArray();
-        for (int idx = 0; idx < buf.length; ++idx)
-            buf[idx] = symbols[random.nextInt(symbols.length)];
-        return new String(buf);
-    }
 }
